Add files via upload
Our project, Retro Games, is an online gaming platform that holds a special place in my heart. As a team of passionate gamers and web developers, we have come together to create an appealing and interactive website that offers a nostalgic experience for users. We wanted to bring back the joy and memories of the classic retro games that were popular in the early 2000s.

The main objective of Retro Games is to provide a touch of nostalgia to the generation that grew up playing these games, as well as introduce the new generation to the magic of retro gaming. Our website features a wide collection of beloved childhood games, including timeless classics like Pac-Man and Pong. We carefully selected games that do not require high memory space or graphic specifications, making them accessible to all PC owners.